검색 조회, 통합 , 검색 , 업체 2020/11/15 #3

+++ src/main/java/com/tobe/fishking/v2/model/FileDTO.java
@@ -34,7 +34,7 @@ public class FileDTO {
     private LocalDateTime createdDate;
     private LocalDateTime modifiedDate;
     private String fileName;
-    private int pid;
+    private Long pid;
     private String originalFile;
     private String  storedFile;
     private String  thumbnailFile;
